work with media query - web header
