+++ src/hooks/Applier.jsx
@@ -4,13 +4,13 @@ import axios from 'axios';
 const BASE_URI = () => {
     if (process.env.REACT_APP_ENV === 'production') return process.env.REACT_APP_DEV_URI;
     else return process.env.REACT_APP_PROD_URI;
-}
+};
 
 export const useApplierApi = () => {
     const [appliers, setAppliers] = useState({ title: '', introduction: '', questions: [] });
 
     const fetchApplier = async () => {
-        const data = await axios.get(BASE_URI + '/applies/appliers');
+        const data = await axios.get(BASE_URI() + '/applies/appliers');
         setAppliers(data.data.data[0]);
     };
